refactor: move `src/cli` to `src` with prefix `cli`
